feat(dialog): move folder and rename root dialog component
